requirements.txt
Flask==2.3.2
redis==4.6.0
rq==1.13.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.21
